deleteCart && showCart not 100% working
body is not set

+++ backend/routes/cart.js
@@ -56,8 +56,9 @@ router.put("/:id", verifyTokenAndAuthorization, async (req, res) => {
 
 //DELETE
 router.delete("/:id", verifyTokenAndAuthorization, async (req, res) => {
+    console.log("userid " + req.params.id)
     try {
-        await Cart.findByIdAndDelete(req.params.id);
+        await Cart.findByIdAndDelete({ userId: req.params.id });
         res.status(200).json("Cart has been deleted...");
     } catch (err) {
         res.status(500).json(err);
@@ -65,7 +66,8 @@ router.delete("/:id", verifyTokenAndAuthorization, async (req, res) => {
 });
 
 //GET USER CART
-router.get("/getCart", verifyTokenAndAuthorization, async (req, res) => {
+router.get("/find/:userId", async (req, res) => {
+    console.log("find/user " + JSON.stringify(req.body));
     try {
         const cart = await Cart.findOne({ userId: req.params.userId });
         res.status(200).json(cart);
